2-add

+++ 0x12-singly_linked_lists/2-add_node.c
@@ -9,7 +9,7 @@
 list_t *add_node(list_t **head, const char *str)
 {
 	list_t *target; /* declaring pointer to new node */
-	int i, count = 0;
+	size_t i, count = 0;
 
 	if (head == NULL)
 		return (NULL);
